intermediate code for first time simulation

+++ headers/Utility.h
@@ -2,6 +2,6 @@
 
 int uniformDistributionInt(int min, int max);
 
-double uniformDistributionDouble(int min, int max);
+double uniformDistributionDouble(double min, double max);
 
 double exponentialDistribution(double lambda);
